+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Support for ORDER BY with LIMIT
+  ([#954](https://github.com/feldera/feldera/pull/954))
 - Allow configuring resource requirements per pipeline
   ([940](https://github.com/feldera/feldera/pull/940))
 - Restructure and expansion of cloud documentation
@@ -36,8 +38,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support for ORDER BY with LIMIT
-  ([#954](https://github.com/feldera/feldera/pull/954))
 - Preliminary support for BINARY and VARBINARY SQL
   data types
   ([#917](https://github.com/feldera/feldera/pull/917))